refactor: SideMenu 및 관련 AccordionMenu 컴포넌트를 widget으로 이동
