+++ src/widget/sideMenu/AccordionMenu.tsx
